+++ mayan/apps/document_states/api_views/workflow_template_api_views.py
@@ -16,8 +16,8 @@ from ..permissions import (
 from ..serializers import (
     WorkflowTemplateDocumentTypeAddSerializer,
     WorkflowTemplateDocumentTypeRemoveSerializer, WorkflowTemplateSerializer,
-    WorkflowTemplateStateSerializer, WorkflowTemplateTransitionSerializer,
-    WorkflowTransitionFieldSerializer
+    WorkflowTemplateStateActionSerializer, WorkflowTemplateStateSerializer,
+    WorkflowTemplateTransitionSerializer, WorkflowTransitionFieldSerializer
 )
 
 
@@ -198,6 +198,94 @@ class APIWorkflowTemplateStateView(
 
     def get_queryset(self):
         return self.external_object.states.all()
+
+
+class APIWorkflowTemplateStateActionListView(generics.ListCreateAPIView):
+    """
+    get: Returns a list of all the workflow template state actions.
+    post: Create a new workflow template state action.
+    """
+    ordering_fields = ('label', 'enabled', 'id')
+    serializer_class = WorkflowTemplateStateActionSerializer
+
+    def get_instance_extra_data(self):
+        return {
+            '_event_actor': self.request.user,
+            'state': self.get_workflow_template_state()
+        }
+
+    def get_queryset(self):
+        return self.get_workflow_template_state().actions.all()
+
+    def get_workflow_template_state(self):
+        if self.request.method == 'GET':
+            permission_required = permission_workflow_template_view
+        else:
+            permission_required = permission_workflow_template_edit
+
+        queryset = AccessControlList.objects.restrict_queryset(
+            permission=permission_required,
+            queryset=self.get_workflow_template().states.all(),
+            user=self.request.user
+        )
+
+        return get_object_or_404(
+            klass=queryset, pk=self.kwargs['workflow_template_state_id']
+        )
+
+    def get_workflow_template(self):
+        return get_object_or_404(
+            klass=Workflow.objects.all(),
+            pk=self.kwargs['workflow_template_id']
+        )
+
+
+class APIWorkflowTemplateStateActionDetailView(
+    generics.RetrieveUpdateDestroyAPIView
+):
+    """
+    delete: Delete the selected workflow template state action.
+    get: Return the details of the selected workflow template state action.
+    patch: Edit the selected workflow template state action.
+    put: Edit the selected workflow template state action.
+    """
+    mayan_object_permissions = {
+        'DELETE': (permission_workflow_template_edit,),
+        'GET': (permission_workflow_template_view,),
+        'PATCH': (permission_workflow_template_edit,),
+        'PUT': (permission_workflow_template_edit,),
+    }
+    lookup_url_kwarg = 'workflow_template_state_action_id'
+    serializer_class = WorkflowTemplateStateActionSerializer
+
+    def get_instance_extra_data(self):
+        return {
+            '_event_actor': self.request.user,
+        }
+
+    def get_queryset(self):
+        return self.get_workflow_template_state().actions.all()
+
+    def get_workflow_template_state(self):
+        if self.request.method == 'GET':
+            permission_required = permission_workflow_template_view
+        else:
+            permission_required = permission_workflow_template_edit
+
+        queryset = AccessControlList.objects.restrict_queryset(
+            permission=permission_required,
+            queryset=self.get_workflow_template().states.all(),
+            user=self.request.user
+        )
+        return get_object_or_404(
+            klass=queryset, pk=self.kwargs['workflow_template_state_id']
+        )
+
+    def get_workflow_template(self):
+        return get_object_or_404(
+            klass=Workflow.objects.all(),
+            pk=self.kwargs['workflow_template_id']
+        )
 
 
 # Workflow transition views